+++ bindings/java/native/src/classes/message.rs
@@ -7,7 +7,7 @@ use iota_wallet::{
     address::{AddressWrapper, OutputKind},
     message::{
         Message as MessageRust, BlockId, RemainderValueStrategy as RemainderValueStrategyRust,
-        Transfer as TransferRust, TransferBuilder as TransferBuilderRust,
+        Transaction as TransactionRust, TransactionBuilder as TransactionBuilderRust,
     },
 };
 
@@ -28,28 +28,28 @@ pub fn remainder_type_enum_to_type(strategy: RemainderValueStrategy) -> Remainde
     }
 }
 
-pub struct Transfer {
-    transfer: TransferRust,
+pub struct Transaction {
+    transaction: TransactionRust,
 }
 
-impl Transfer {
-    pub fn to_inner(self) -> TransferRust {
-        self.transfer
+impl Transaction {
+    pub fn to_inner(self) -> TransactionRust {
+        self.transaction
     }
 
-    pub fn builder(address: AddressWrapper, amount: u64, output_kind: Option<OutputKind>) -> TransferBuilder {
-        TransferBuilder::new(address, amount, output_kind)
+    pub fn builder(address: AddressWrapper, amount: u64, output_kind: Option<OutputKind>) -> TransactionBuilder {
+        TransactionBuilder::new(address, amount, output_kind)
     }
 }
 
-pub struct TransferBuilder {
-    builder: Rc<RefCell<Option<TransferBuilderRust>>>,
+pub struct TransactionBuilder {
+    builder: Rc<RefCell<Option<TransactionBuilderRust>>>,
 }
 
-impl TransferBuilder {
+impl TransactionBuilder {
     pub fn new(address: AddressWrapper, amount: u64, output_kind: Option<OutputKind>) -> Self {
         Self {
-            builder: Rc::new(RefCell::new(Option::from(TransferBuilderRust::new(
+            builder: Rc::new(RefCell::new(Option::from(TransactionBuilderRust::new(
                 address,
                 NonZeroU64::new(amount).unwrap(),
                 output_kind,
@@ -57,7 +57,7 @@ impl TransferBuilder {
         }
     }
 
-    pub fn new_with_builder(builder: TransferBuilderRust) -> Self {
+    pub fn new_with_builder(builder: TransactionBuilderRust) -> Self {
         Self {
             builder: Rc::new(RefCell::new(Option::from(builder))),
         }
@@ -70,7 +70,7 @@ impl TransferBuilder {
             .take()
             .unwrap()
             .with_remainder_value_strategy(remainder_type_enum_to_type(strategy));
-        TransferBuilder::new_with_builder(new_builder)
+        TransactionBuilder::new_with_builder(new_builder)
     }
 
     pub fn with_remainder_to_account_with_address(&mut self, address: AddressWrapper) -> Self {
@@ -80,7 +80,7 @@ impl TransferBuilder {
             .take()
             .unwrap()
             .with_remainder_value_strategy(RemainderValueStrategyRust::AccountAddress(address));
-        TransferBuilder::new_with_builder(new_builder)
+        TransactionBuilder::new_with_builder(new_builder)
     }
 
     pub fn with_indexation(&mut self, indexation: IndexationPayload) -> Self {
@@ -90,18 +90,18 @@ impl TransferBuilder {
             .take()
             .unwrap()
             .with_indexation(indexation.to_inner());
-        TransferBuilder::new_with_builder(new_builder)
+        TransactionBuilder::new_with_builder(new_builder)
     }
 
     pub fn with_skip_sync(&mut self) -> Self {
         let new_builder = self.builder.borrow_mut().take().unwrap().with_skip_sync();
-        TransferBuilder::new_with_builder(new_builder)
+        TransactionBuilder::new_with_builder(new_builder)
     }
 
-    /// Builds the transfer.
-    pub fn finish(&self) -> Transfer {
-        Transfer {
-            transfer: self.builder.borrow_mut().take().unwrap().finish(),
+    /// Builds the transaction.
+    pub fn finish(&self) -> Transaction {
+        Transaction {
+            transaction: self.builder.borrow_mut().take().unwrap().finish(),
         }
     }
 }