+++ examples/05_transaction.rs
@@ -31,12 +31,12 @@ async fn main() -> Result<()> {
         address: "atoi1qpszqzadsym6wpppd6z037dvlejmjuke7s24hm95s9fg9vpua7vluehe53e".to_string(),
         amount: 1_000_000,
     }];
-    let transfer_result = account.send_amount(outputs, None).await?;
+    let transaction_result = account.send_amount(outputs, None).await?;
 
     println!(
         "Transaction: {} Block sent: http://localhost:14265/api/v2/blocks/{}",
-        transfer_result.transaction_id,
-        transfer_result.block_id.expect("No block created yet")
+        transaction_result.transaction_id,
+        transaction_result.block_id.expect("No block created yet")
     );
 
     Ok(())