+++ src/account/operations/transaction/build_transaction.rs
@@ -18,18 +18,18 @@ use iota_client::{
     secret::types::InputSigningData,
 };
 
-use crate::account::{handle::AccountHandle, operations::transfer::TransferOptions};
+use crate::account::{handle::AccountHandle, operations::transaction::TransactionOptions};
 #[cfg(feature = "events")]
-use crate::events::types::{TransferProgressEvent, WalletEvent};
+use crate::events::types::{TransactionProgressEvent, WalletEvent};
 
 impl AccountHandle {
     /// Function to build the transaction essence from the selected in and outputs
     pub(crate) async fn build_transaction_essence(
         &self,
         selected_transaction_data: SelectedTransactionData,
-        options: Option<TransferOptions>,
+        options: Option<TransactionOptions>,
     ) -> crate::Result<PreparedTransactionData> {
-        log::debug!("[TRANSFER] build_transaction");
+        log::debug!("[TRANSACTION] build_transaction");
         let build_transaction_essence_start_time = Instant::now();
 
         let mut inputs_for_essence: Vec<Input> = Vec::new();
@@ -86,13 +86,13 @@ impl AccountHandle {
             let account_index = self.read().await.index;
             self.event_emitter.lock().await.emit(
                 account_index,
-                WalletEvent::TransferProgress(TransferProgressEvent::PreparedTransaction(Box::new(
+                WalletEvent::TransactionProgress(TransactionProgressEvent::PreparedTransaction(Box::new(
                     PreparedTransactionDataDto::from(&prepared_transaction_data),
                 ))),
             );
         }
         log::debug!(
-            "[TRANSFER] finished build_transaction in {:.2?}",
+            "[TRANSACTION] finished build_transaction in {:.2?}",
             build_transaction_essence_start_time.elapsed()
         );
         Ok(prepared_transaction_data)