+++ src/account/operations/transaction/mod.rs
@@ -1,8 +1,6 @@
 // Copyright 2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-// transfer or transaction?
-
 mod build_transaction;
 pub(crate) mod high_level;
 mod input_selection;
@@ -25,9 +23,9 @@ use iota_client::{
 };
 use serde::Serialize;
 
-pub use self::options::{RemainderValueStrategy, TransferOptions};
+pub use self::options::{RemainderValueStrategy, TransactionOptions};
 #[cfg(feature = "events")]
-use crate::events::types::{TransferProgressEvent, WalletEvent};
+use crate::events::types::{TransactionProgressEvent, WalletEvent};
 use crate::{
     account::{
         handle::AccountHandle,
@@ -37,9 +35,9 @@ use crate::{
     iota_client::Error,
 };
 
-/// The result of a transfer, block_id is an option because submitting the transaction could fail
+/// The result of a transaction, block_id is an option because submitting the transaction could fail
 #[derive(Debug, Serialize)]
-pub struct TransferResult {
+pub struct TransactionResult {
     #[serde(rename = "transactionId")]
     pub transaction_id: TransactionId,
     #[serde(rename = "blockId")]
@@ -50,8 +48,9 @@ impl AccountHandle {
     /// Send a transaction, if sending a block fails, the function will return None for the block_id, but the wallet
     /// will retry sending the transaction during syncing.
     /// ```ignore
-    /// let outputs = vec![TransferOutput {
-    ///     address: "atoi1qpszqzadsym6wpppd6z037dvlejmjuke7s24hm95s9fg9vpua7vluehe53e".to_string(),
+    /// let outputs = vec![TransactionOutput {
+    ///     address: "// transaction or transaction?
+    /// atoi1qpszqzadsym6wpppd6z037dvlejmjuke7s24hm95s9fg9vpua7vluehe53e".to_string(),
     ///     amount: 1_000_000,
     ///     output_kind: None,
     /// }];
@@ -59,7 +58,7 @@ impl AccountHandle {
     /// let res = account_handle
     ///     .send(
     ///         outputs,
-    ///         Some(TransferOptions {
+    ///         Some(TransactionOptions {
     ///             remainder_value_strategy: RemainderValueStrategy::ReuseAddress,
     ///             ..Default::default()
     ///         }),
@@ -70,7 +69,11 @@ impl AccountHandle {
     ///     println!("Block sent: {}", block_id);
     /// }
     /// ```
-    pub async fn send(&self, outputs: Vec<Output>, options: Option<TransferOptions>) -> crate::Result<TransferResult> {
+    pub async fn send(
+        &self,
+        outputs: Vec<Output>,
+        options: Option<TransactionOptions>,
+    ) -> crate::Result<TransactionResult> {
         // here to check before syncing, how to prevent duplicated verification (also in prepare_transaction())?
         // Checking it also here is good to return earlier if something is invalid
         let byte_cost_config = self.client.get_byte_cost_config().await?;
@@ -85,7 +88,7 @@ impl AccountHandle {
             let account_index = self.read().await.index;
             self.event_emitter.lock().await.emit(
                 account_index,
-                WalletEvent::TransferProgress(TransferProgressEvent::SyncingAccount),
+                WalletEvent::TransactionProgress(TransactionProgressEvent::SyncingAccount),
             );
         }
         if !options.clone().unwrap_or_default().skip_sync {
@@ -95,29 +98,29 @@ impl AccountHandle {
             }))
             .await?;
         }
-        self.finish_transfer(outputs, options).await
+        self.finish_transaction(outputs, options).await
     }
 
     /// Separated function from send, so syncing isn't called recursively with the consolidation function, which sends
-    /// transfers
-    pub async fn finish_transfer(
+    /// transactions
+    pub async fn finish_transaction(
         &self,
         outputs: Vec<Output>,
-        options: Option<TransferOptions>,
-    ) -> crate::Result<TransferResult> {
-        log::debug!("[TRANSFER] finish_transfer");
+        options: Option<TransactionOptions>,
+    ) -> crate::Result<TransactionResult> {
+        log::debug!("[TRANSACTION] finish_transaction");
 
         let prepared_transaction_data = self.prepare_transaction(outputs, options).await?;
 
-        self.sign_and_submit_transfer(prepared_transaction_data).await
+        self.sign_and_submit_transaction(prepared_transaction_data).await
     }
 
     /// Sign a transaction, submit it to a node and store it in the account
-    pub async fn sign_and_submit_transfer(
+    pub async fn sign_and_submit_transaction(
         &self,
         prepared_transaction_data: PreparedTransactionData,
-    ) -> crate::Result<TransferResult> {
-        log::debug!("[TRANSFER] sign_and_submit_transfer");
+    ) -> crate::Result<TransactionResult> {
+        log::debug!("[TRANSACTION] sign_and_submit_transaction");
 
         let signed_transaction_data = match self.sign_transaction_essence(&prepared_transaction_data).await {
             Ok(res) => res,
@@ -135,16 +138,16 @@ impl AccountHandle {
     pub async fn sync_and_prepare_transaction(
         &self,
         outputs: Vec<Output>,
-        options: Option<TransferOptions>,
+        options: Option<TransactionOptions>,
     ) -> crate::Result<PreparedTransactionData> {
-        log::debug!("[TRANSFER] sync_and_prepare_transaction");
+        log::debug!("[TRANSACTION] sync_and_prepare_transaction");
         // sync account before sending a transaction
         #[cfg(feature = "events")]
         {
             let account_index = self.read().await.index;
             self.event_emitter.lock().await.emit(
                 account_index,
-                WalletEvent::TransferProgress(TransferProgressEvent::SyncingAccount),
+                WalletEvent::TransactionProgress(TransactionProgressEvent::SyncingAccount),
             );
         }
         if !options.clone().unwrap_or_default().skip_sync {
@@ -162,8 +165,8 @@ impl AccountHandle {
     pub async fn submit_and_store_transaction(
         &self,
         signed_transaction_data: SignedTransactionData,
-    ) -> crate::Result<TransferResult> {
-        log::debug!("[TRANSFER] submit_and_store_transaction");
+    ) -> crate::Result<TransactionResult> {
+        log::debug!("[TRANSACTION] submit_and_store_transaction");
 
         // Validate transaction before sending and storing it
         let (local_time, milestone_index) = self.client.get_time_and_milestone_checked().await?;
@@ -212,10 +215,10 @@ impl AccountHandle {
         account.pending_transactions.insert(transaction_id);
         #[cfg(feature = "storage")]
         {
-            log::debug!("[TRANSFER] storing account {}", account.index());
+            log::debug!("[TRANSACTION] storing account {}", account.index());
             self.save(Some(&account)).await?;
         }
-        Ok(TransferResult {
+        Ok(TransactionResult {
             transaction_id,
             block_id,
         })