+++ src/lib.rs
@@ -34,7 +34,7 @@ pub use iota_client::{self, secret};
 pub use primitive_types::U256;
 
 pub use self::{
-    account::operations::transfer::high_level::{
+    account::operations::transaction::high_level::{
         mint_native_token::NativeTokenOptions, mint_nfts::NftOptions, send_amount::AddressWithAmount,
         send_micro_transaction::AddressWithMicroAmount, send_native_tokens::AddressNativeTokens,
         send_nft::AddressAndNftId,